Delegating GAM Payments to Non-Profit Square accounts (Think!Chinatown) (#207)
To fully facilitate payments, square card nonce's have to be encoded with the right application id--so for appropriate transactions, we need to use the Think!Chinatown id. The Square React component requires this to be passed in as a prop, s owe need to have it in the right context. This involved piping the `non_profit_location_id` down from the initial seller call to not have to make a second call for this info.

Overview:
- Seller now has `non_profit_location_id` as a property, so handling that, then passing it down through `props` down to the `SquareModal`
- Checking if Location ID's match, and if so, switching to the appropriate credential set

Supporting Details:
- Adding all Think!Chinatown credentials to the env
- Swapped the default sandbox square location ID to our sandbox app
- Ran Prettier on the files changed and also fixed some typos

* Switching app and location id's for think chinatown

* passing the location id through props rather than calling it

* removing comment

* formatting

* switching local to point back to staging

* updating .env.development's local backend swap to be on two lines