Merchant Voucher Dashboard (#222)
Here's a good first pass at the dashboard for merchants to see all the vouchers for their business in real time. Right now we're working with the idea that @justintmckibben and I came up with to have a unique URL that the merchants can access. He put in the back-end support for that here: sendchinatownlove/ruby#175

I tried to follow the Figma designs as much as possible, but we're not returning quite all the data that is laid out, and we don't have any support for interactivity through the api right now. I think this is good for now, and we can revisit those features if/when Vouchers become more important for us.

I did some searching for table frameworks, and there a lot of really powerful ones with tons of customization, but react-fiterable-table seemed to do pretty much everything we wanted out of the box. It hasn't been updated for about a year and has some console logs about deprecated methods, but I think that's ok for now. It expected bootstrap css, so I've added that for this page and done a bunch of custom styling on top of it. Again, we can revisit if/when vouchers become more important. For now it gives us really easy sorting, filtering, and pagination!

The page is pretty responsive at widths ~700px and wider. Unfortunately the table doesn't really work that will when narrow. It should work well on a tablet.

* Super minimal proof of concept of page

- Piped page and params
- Header

* metadata header done

* working POC with react-filterable-table

* Adding bootstrap and lots of additional styling

* Getting seller info; reusing data call; error page

- Getting the seller info--no more mocking text
- Reusing the one data call and passing to the table, no more dataEndpoint prop
- Displaying the error (404) page on server error now, not the best but ok
- Small cleanup and fixes
- Verified working for no gift cards as well

* Updating url to seller_id/dashboard/secret_id

- Minor refactoring

* Adding bilingual text and styling

* Rename MerchantGiftCardsPage to MerchantVoucherDashboard

* Reorganizing and refactoring